Continued to clean up the unit tests.
Moved the Point tests from test_predicates to test_primitives.
Uncommented all the predicate tests, which now fail, of course.
